+++ src/main/kotlin/dev/nybroe/collector/Util.kt
@@ -1,0 +1,7 @@
+package dev.nybroe.collector
+
+import com.jetbrains.php.lang.psi.elements.FunctionReference
+
+fun FunctionReference.isGlobalFunctionCallWithName(name: String): Boolean {
+    return this.name == name
+}