+++ src/test/resources/inspections/ArrayMapToCollectionInspection/array_map-callback-and-array.php
@@ -1,0 +1,5 @@
+<?php
+
+$b = <caret>array_map(function($n) {
+    return ($n * $n * $n);
+}, [1, 2, 3, 4, 5]);