+++ src/test/resources/inspections/ArrayMapToCollectionInspection/array_map-callback-and-variable.php
@@ -1,0 +1,6 @@
+<?php
+
+$a = [1, 2, 3, 4, 5];
+$b = <caret>array_map(function($n) {
+    return ($n * $n * $n);
+}, $a);