+++ src/test/resources/inspections/ArrayMapToCollectionInspection/namespaced_array_map.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace MyAwesomeName\Good;
+
+$a = [1, 2, 3, 4, 5];
+$b = collect($a)->map(function ($n) {
+    return ($n * $n * $n);
+})->all();