+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 ### Added
 - Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
 - Added foreach to collection refactoring
+- Added array map to collection refactoring